feat: java implementation
