+++ Sudoku.ChocoSolver/ChocoSolver.cs
@@ -1,57 +1,17 @@
-﻿using Python.Runtime;
-using Sudoku.ChocoSolver;
-using Sudoku.Shared;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Resources;
-using System.Text;
-using System.Threading.Tasks;
+﻿using Sudoku.Shared;
+using fr.epita;
 
-namespace Sudoku.Demo
+namespace Sudoku.ChocoSolver;
+
+public class ChocoSolverBase : ISudokuSolver
 {
-    public class ChocoSolver : PythonSolverBase
+	public SudokuGrid Solve(SudokuGrid s)
+	{
+		var toSolve = s.CloneSudoku();
+		var javaSolver = new fr.epita.SudokuSolver(toSolve.Cells);
 
-    {
-        public override Shared.SudokuGrid Solve(Shared.SudokuGrid s)
-        {
-            //System.Diagnostics.Debugger.Break();
+		javaSolver.solve();
 
-            //For some reason, the Benchmark runner won't manage to get the mutex whereas individual execution doesn't cause issues
-            using (Py.GIL())
-            {
-                // create a Python scope
-                using (PyModule scope = Py.CreateScope())
-                {
-                    // convert the Cells array object to a PyObject
-                    PyObject pyCells = s.Cells.ToPython();
-
-                    // create a Python variable "instance"
-                    scope.Set("instance", pyCells);
-
-                    //this.AddNumpyConverterScript(scope);
-
-                    // run the Python script
-                    string code = Resource1.solver_py;
-                    scope.Exec(code);
-
-                    //Retrieve solved Sudoku variable
-                    var result = scope.Get("r");
-
-                    //Convert back to C# object
-                    var managedResult = result.As<int[][]>();
-                    //var convertesdResult = managedResult.Select(objList => objList.Select(o => (int)o).ToArray()).ToArray();
-                    return new Shared.SudokuGrid() { Cells = managedResult };
-                }
-            }
-        }
-        
-        protected override void InitializePythonComponents()
-        {
-            //declare your pip packages here
-            InstallPipModule("numpy");
-            InstallPipModule("pycsp3");
-            base.InitializePythonComponents();
-        }
-    }
+		return toSolve;
+	}
 }
