+++ Sudoku.ChocoSolver/Sudoku.ChocoSolver.csproj
@@ -25,11 +25,16 @@
     </EmbeddedResource>
   </ItemGroup>
 
-  <!-- <ItemGroup>
-    <PackageReference Include="IKVM" Version="8.5.0-develop0690" />
-  </ItemGroup>
-  
   <ItemGroup>
-    <IkvmReference Include="SudokuSolver-1.jar" />
-  </ItemGroup> -->
+    <PackageReference Include="IKVM" Version="8.4.5" />
+    <PackageReference Include="IKVM.Image.JDK" Version="8.4.5" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <IkvmReference Include=".\ChocoSolverJava\SudokuSolver.jar">
+      <AssemblyName>Sudoku.ChocoSolverJVM</AssemblyName>
+      <AssemblyVersion>3.2.1.0</AssemblyVersion>
+      <AssemblyFileVersion>3.0.0.0</AssemblyFileVersion>
+    </IkvmReference>
+  </ItemGroup>
 </Project>
